Also move walsim
