+++ .gitignore
@@ -68,5 +68,5 @@ sei-db/bench/cryptosim/data/**
 sei-db/bench/cryptosim/bin/
 sei-db/bench/cryptosim/logs/
 sei-db/bench/blocksim/bin/
-sei-db/seiwal/walsim/bin/
+sei-db/bench/walsim/bin/
 sei-db/db_engine/litt/bin/
