+++ sei-db/bench/walsim/cmd/configure-logger/main.go
@@ -13,8 +13,8 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/sei-protocol/sei-chain/sei-db/bench/walsim"
 	"github.com/sei-protocol/sei-chain/sei-db/common/utils"
-	"github.com/sei-protocol/sei-chain/sei-db/seiwal/walsim"
 )
 
 func main() {
