+++ sei-db/bench/walsim/cmd/walsim/main.go
@@ -7,9 +7,9 @@ import (
 	"os"
 	"os/signal"
 
+	"github.com/sei-protocol/sei-chain/sei-db/bench/walsim"
 	"github.com/sei-protocol/sei-chain/sei-db/common/metrics"
 	"github.com/sei-protocol/sei-chain/sei-db/common/utils"
-	"github.com/sei-protocol/sei-chain/sei-db/seiwal/walsim"
 )
 
 func main() {
