+++ sei-db/bench/walsim/config/standard.json
